requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
api-connectors==3.1
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
behave==1.2.6
binaryornot==0.4.4
blinker==1.7.0
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chromedriver==2.24.1
chromedriver-binary==126.0.6465.2.0
chromedriver-binary-auto==0.3.1
cleo==2.1.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
configparser==7.0.0
contextlib2==21.6.0
contourpy==1.2.0
cookiecutter==2.5.0
crashtest==0.4.1
crayons==0.4.0
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.5.14
datasets==2.16.1
DateTime==5.4
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
distlib==0.3.8
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.1
django-crispy-forms==2.1
django-rest-action-permissions==2.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.2.0
env==0.1.0
et-xmlfile==1.1.0
execnet==2.0.2
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flet==0.18.0
flet-core==0.18.0
flet-runtime==0.18.0
fonttools==4.49.0
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2023.10.0
furl==2.1.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.15.0
google-api-python-client==2.113.0
google-api-python-client-py3==1.2
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gptcache==0.1.43
greenlet==3.0.3
gspread==5.12.4
gTTS==2.5.1
h11==0.14.0
h3==3.7.7
hashring==1.5.1
html5lib==1.1
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.1
httpx==0.24.1
huggingface-hub==0.20.3
idna==2.10
importlib-metadata==6.11.0
iniconfig==2.0.0
installer==0.7.0
integrate==1.3.0
ipykernel==6.28.0
ipython==8.19.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.0
jupyterlab_widgets==3.0.10
keyring==24.3.1
kiwisolver==1.4.5
langchain-community==0.0.13
langchain-core==0.1.13
langsmith==0.0.83
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
memcache==0.12.0
mock==5.1.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==8.2.0
mysqlclient==2.2.1
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
numexpr==2.8.8
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.2.2
objects==0.3.1
openapi-schema-pydantic==1.2.4
opencv-python==4.9.0.80
openpyxl==3.1.2
orderedmultidict==1.0.1
orjson==3.10.6
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
parse==1.20.2
parse-type==0.6.2
parso==0.8.3
path==16.10.0
path.py==12.5.0
pdf2image==1.17.0
pdfkit==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.1
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.1.0
playsound==1.3.0
plotly==5.19.0
pluggy==1.3.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prompt-toolkit==3.0.43
protobuf==4.21.12
psutil==5.9.7
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-openaq==1.1.0
py3-wkhtmltopdf==0.4.1
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.14.6
pydeck==0.8.1b0
pydicom==2.4.4
pydub==0.25.1
pygame==2.5.2
Pygments==2.17.2
PyHamcrest==2.1.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.1
pypdf==4.0.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pypng==0.20220715.0
pyproject_hooks==1.1.0
PySocks==1.7.1
pytest==7.4.0
pytest-shutil==1.7.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.1
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
rapidfuzz==3.9.0
readchar==4.0.5
referencing==0.32.1
repath==0.9.0
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
seaborn==0.13.2
selenium==4.22.0
setuptools==69.0.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.29
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
streamlit==1.31.1
sympy==1.12
tabula-py==2.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
timezonefinder==6.4.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.5
toolz==0.12.0
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
trio==0.25.0
trio-websocket==0.11.1
trove-classifiers==2024.4.10
ttkthemes==3.2.2
twilio==8.11.0
typer==0.12.3
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.18
urllib3-secure-extra==0.1.0
uvicorn==0.30.1
validators==0.22.0
virtualenv==20.25.1
watchdog==3.0.0
watchfiles==0.22.0
wcwidth==0.2.12
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.1
wheel==0.42.0
widgetsnbextension==4.0.10
windows-curses==2.3.2
wsproto==1.2.0
XlsxWriter==3.2.0
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
zope.interface==6.1
